Added caps lock support
Bit unusual: caps lock is managed totally by the MCU. If caps lock
is being turned off, then send a key up scancode for the caps lock
key. If it is going on, then send a key down. The actual button being
released is bsically ignored.

This means the MPU code can track the key just like it can an ordinary
shift key.