Added key debouncing
A little inefficent; a byte size counter is maintained for each key.
This counter maxs out at around 5, so really either 4 keys should be
stuffed into a byte or perhaps 2 keys (maxing the counter at 16). But
there's still enough RAM in the AVR (512 bytes) so 128 bytes for the
debouncing counters is fine, for now.

This counter is reset to 0 on startup, meaning it is "off". On a key
event it is set to 1, which then counts up to a trip point, after
which the event is generated and the counter reset to 0. Seems to
work pretty well.

Some comment cleanups.